backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-CORS
Werkzeug
psycopg2
PyMuPDF
Pillow
gtts
speechrecognition
pydub
googletrans==4.0.0-rc1
python-dotenv
pdfrw
reportlab
PyPDF2
FPDF
pikepdf